requirements.txt
alembic==1.4.2
astroid==2.4.1
autopep8==1.5.2
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
Flask==1.1.2
Flask-API==2.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
http-ece==1.1.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
py-vapid==1.7.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pywebpush==1.11.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
